refactor(create): route wizard through token picker bridge
Closes #607
